+++ categories.go
@@ -4,28 +4,25 @@ import (
 	"context"
 	"net/http"
 	"net/url"
-	"strconv"
 
 	"github.com/henrikah/kick-go-sdk/internal/endpoints"
 	"github.com/henrikah/kick-go-sdk/kickapitypes"
 	"github.com/henrikah/kick-go-sdk/kickcontracts"
 	"github.com/henrikah/kick-go-sdk/kickerrors"
+	"github.com/henrikah/kick-go-sdk/kickfilters"
 )
 
 type categoryClient struct {
 	client *apiClient
 }
 
-func newCategoryClient(client *apiClient) kickcontracts.Category {
+func newCategoryService(client *apiClient) kickcontracts.Category {
 	return &categoryClient{
 		client: client,
 	}
 }
-func (c *categoryClient) SearchCategories(ctx context.Context, accessToken string, searchQuery string, pageNumber int) (*kickapitypes.GetCategoriesResponse, error) {
+func (c *categoryClient) SearchCategories(ctx context.Context, accessToken string, filters kickfilters.CategoriesFilter) (*kickapitypes.GetCategoriesResponse, error) {
 	if err := kickerrors.ValidateAccessToken(accessToken); err != nil {
-		return nil, err
-	}
-	if err := kickerrors.ValidatePageNumber(pageNumber); err != nil {
 		return nil, err
 	}
 
@@ -34,38 +31,20 @@ func (c *categoryClient) SearchCategories(ctx context.Context, accessToken strin
 		return nil, err
 	}
 
-	queryParams := categoriesURL.Query()
-	queryParams.Set("q", searchQuery)
-	queryParams.Set("page", strconv.Itoa(pageNumber))
+	if filters != nil {
+		queryParams, filterErr := filters.ToQueryString()
+		if filterErr != nil {
+			return nil, filterErr
+		}
 
-	categoriesURL.RawQuery = queryParams.Encode()
+		categoriesURL.RawQuery = queryParams.Encode()
+	}
 
 	var categoriesData kickapitypes.GetCategoriesResponse
 
-	if err = c.client.makeJSONRequest(ctx, http.MethodGet, categoriesURL.String(), nil, &accessToken, &categoriesData); err != nil {
+	if err = c.client.requester.MakeJSONRequest(ctx, http.MethodGet, categoriesURL.String(), nil, &accessToken, &categoriesData); err != nil {
 		return nil, err
 	}
 
 	return &categoriesData, nil
-}
-func (c *categoryClient) GetCategoryByCategoryID(ctx context.Context, accessToken string, categoryID int) (*kickapitypes.GetCategoryResponse, error) {
-	if err := kickerrors.ValidateAccessToken(accessToken); err != nil {
-		return nil, err
-	}
-	if err := kickerrors.ValidateCategoryID(categoryID); err != nil {
-		return nil, err
-	}
-
-	categoriesURL, err := url.Parse(endpoints.ViewCategoryDetailsURL(categoryID))
-	if err != nil {
-		return nil, err
-	}
-
-	var categoryData kickapitypes.GetCategoryResponse
-
-	if err := c.client.makeJSONRequest(ctx, http.MethodGet, categoriesURL.String(), nil, &accessToken, &categoryData); err != nil {
-		return nil, err
-	}
-
-	return &categoryData, nil
 }