+++ kickcontracts/event_subscription.go
@@ -16,16 +16,21 @@ type EventsSubscription interface {
 	//
 	// Example:
 	//
-	//	client, err := kick.NewAPIClient(kickapitypes.APIClientConfig{...})
+	//	client, err := kick.NewAPIClient(kickapitypes.APIClientConfig{
+	//	    HTTPClient: http.DefaultClient,
+	//	})
 	//	if err != nil {
-	//	    log.Printf("could not create APIClient: %v", err)
-	//	    return nil, err
+	//	    log.Fatal(err)
 	//	}
 	//
 	//	subscriptions, err := client.EventsSubscription().GetEventSubscriptions(context.TODO(), accessToken)
 	//	if err != nil {
-	//	    log.Printf("could not get event subscriptions: %v", err)
-	//	    return nil, err
+	//		var apiErr *kickerrors.APIError
+	//		if errors.As(err, &apiErr) {
+	//			log.Printf("API error: %d %s", apiErr.StatusCode, apiErr.Message)
+	//		} else {
+	//			log.Printf("internal error: %v", err)
+	//		}
 	//	}
 	GetEventSubscriptions(ctx context.Context, accessToken string) (*kickapitypes.EventSubscription, error)
 
@@ -33,11 +38,22 @@ type EventsSubscription interface {
 	//
 	// Example:
 	//
+	//	client, err := kick.NewAPIClient(kickapitypes.APIClientConfig{
+	//	    HTTPClient: http.DefaultClient,
+	//	})
+	//	if err != nil {
+	//	    log.Fatal(err)
+	//	}
+	//
 	//	events := []kickwebhookenum.WebhookType{kickwebhookenum.ChatMessageSent}
 	//	createEventSubscriptionsResponse, err := client.EventsSubscription().CreateEventSubscriptions(context.TODO(), accessToken, events)
 	//	if err != nil {
-	//	    log.Printf("could not create event subscriptions: %v", err)
-	//	    return nil, err
+	//		var apiErr *kickerrors.APIError
+	//		if errors.As(err, &apiErr) {
+	//			log.Printf("API error: %d %s", apiErr.StatusCode, apiErr.Message)
+	//		} else {
+	//			log.Printf("internal error: %v", err)
+	//		}
 	//	}
 	CreateEventSubscriptions(ctx context.Context, accessToken string, events []kickwebhookenum.WebhookType) (*kickapitypes.CreateEventSubscriptionsResponse, error)
 
@@ -47,10 +63,22 @@ type EventsSubscription interface {
 	//
 	// Example:
 	//
+	//
+	//	client, err := kick.NewAPIClient(kickapitypes.APIClientConfig{
+	//	    HTTPClient: http.DefaultClient,
+	//	})
+	//	if err != nil {
+	//	    log.Fatal(err)
+	//	}
+	//
 	//	createEventSubscriptionsResponse, err := client.EventsSubscription().CreateEventSubscriptionsAsApp(context.TODO(), accessToken, 12345, []kickwebhookenum.WebhookType{kickwebhookenum.ChatMessageSent})
 	//	if err != nil {
-	//	    log.Printf("could not create app event subscriptions: %v", err)
-	//	    return nil, err
+	//		var apiErr *kickerrors.APIError
+	//		if errors.As(err, &apiErr) {
+	//			log.Printf("API error: %d %s", apiErr.StatusCode, apiErr.Message)
+	//		} else {
+	//			log.Printf("internal error: %v", err)
+	//		}
 	//	}
 	CreateEventSubscriptionsAsApp(ctx context.Context, accessToken string, broadcasterUserID int, events []kickwebhookenum.WebhookType) (*kickapitypes.CreateEventSubscriptionsResponse, error)
 
@@ -58,10 +86,21 @@ type EventsSubscription interface {
 	//
 	// Example:
 	//
+	//	client, err := kick.NewAPIClient(kickapitypes.APIClientConfig{
+	//	    HTTPClient: http.DefaultClient,
+	//	})
+	//	if err != nil {
+	//	    log.Fatal(err)
+	//	}
+	//
 	//	err := client.EventsSubscription().DeleteEventSubscriptions(context.TODO(), accessToken, []string{"subscription-id-1", "subscription-id-2"})
 	//	if err != nil {
-	//	    log.Printf("could not delete event subscriptions: %v", err)
-	//	    return err
+	//		var apiErr *kickerrors.APIError
+	//		if errors.As(err, &apiErr) {
+	//			log.Printf("API error: %d %s", apiErr.StatusCode, apiErr.Message)
+	//		} else {
+	//			log.Printf("internal error: %v", err)
+	//		}
 	//	}
 	DeleteEventSubscriptions(ctx context.Context, accessToken string, subscriptionIDs []string) error
 }