+++ kickoauthtypes/oauth_client.go
@@ -1,0 +1,11 @@
+package kickoauthtypes
+
+import (
+	"github.com/henrikah/kick-go-sdk/internal/httpclient"
+)
+
+type OAuthClientConfig struct {
+	ClientID     string
+	ClientSecret string
+	HTTPClient   httpclient.ClientInterface
+}