+++ tests/channels/get_channels_by_user_id_test.go
@@ -13,23 +13,20 @@ import (
 
 func Test_GetChannelsByBroadcasterUserIDMissingAccessToken_Error(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
 	httpClient := http.DefaultClient
 
 	accessToken := ""
 	broadcasters := []int64{1, 2}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 	client, _ := kick.NewAPIClient(config)
 
 	var validationError *kickerrors.ValidationError
 
 	// Act
-	channelsData, err := client.Channel().GetChannelsByBroadcasterUserID(ctx, accessToken, broadcasters)
+	channelsData, err := client.Channel().GetChannelsByBroadcasterUserID(t.Context(), accessToken, broadcasters)
 
 	// Assert
 	if channelsData != nil {
@@ -53,8 +50,6 @@ func Test_GetChannelsByIDUnAuthorized_Error(t *testing.T) {
 	// Arrange
 	errorJSON := `{"message": "Invalid request"}`
 
-	ctx := t.Context()
-
 	accessToken := "access-token"
 	broadcasters := []int64{1, 2}
 
@@ -65,15 +60,13 @@ func Test_GetChannelsByIDUnAuthorized_Error(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   mockClient,
+		HTTPClient: mockClient,
 	}
 	client, _ := kick.NewAPIClient(config)
 
 	var apiError *kickerrors.APIError
 	// Act
-	channelsData, err := client.Channel().GetChannelsByBroadcasterUserID(ctx, accessToken, broadcasters)
+	channelsData, err := client.Channel().GetChannelsByBroadcasterUserID(t.Context(), accessToken, broadcasters)
 
 	// Assert
 	if err == nil {
@@ -91,10 +84,6 @@ func Test_GetChannelsByIDUnAuthorized_Error(t *testing.T) {
 
 func Test_GetChannelsByBroadcasterUserID_Success(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
-	clientID := "test-id"
-	clientSecret := "test-secret"
-
 	accessToken := "access-token"
 	broadcasters := []int64{1, 2}
 
@@ -170,16 +159,14 @@ func Test_GetChannelsByBroadcasterUserID_Success(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
 
 	// Act
 
-	channelsData, err := client.Channel().GetChannelsByBroadcasterUserID(ctx, accessToken, broadcasters)
+	channelsData, err := client.Channel().GetChannelsByBroadcasterUserID(t.Context(), accessToken, broadcasters)
 
 	// Assert
 	if channelsData == nil {