+++ tests/channels/get_current_broadcaster_channel_test.go
@@ -11,10 +11,6 @@ import (
 
 func Test_GetCurrentBroadcasterChannel_Success(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
-	clientID := "test-id"
-	clientSecret := "test-secret"
-
 	accessToken := "access-token"
 
 	expectedJSON := `{
@@ -66,16 +62,14 @@ func Test_GetCurrentBroadcasterChannel_Success(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
 
 	// Act
 
-	channelData, err := client.Channel().GetCurrentBroadcasterChannel(ctx, accessToken)
+	channelData, err := client.Channel().GetCurrentBroadcasterChannel(t.Context(), accessToken)
 
 	// Assert
 	if channelData == nil {