+++ tests/channels/update_channel_test.go
@@ -14,7 +14,6 @@ import (
 
 func Test_UpdateChannelMissingAccessToken_Error(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
 	httpClient := http.DefaultClient
 
 	accessToken := ""
@@ -26,9 +25,7 @@ func Test_UpdateChannelMissingAccessToken_Error(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
@@ -36,7 +33,7 @@ func Test_UpdateChannelMissingAccessToken_Error(t *testing.T) {
 	var validationError *kickerrors.ValidationError
 
 	// Act
-	err := client.Channel().UpdateChannel(ctx, accessToken, channelData)
+	err := client.Channel().UpdateChannel(t.Context(), accessToken, channelData)
 
 	// Assert
 	if err == nil {
@@ -66,9 +63,7 @@ func Test_UpdateChannelNegativeCategoryID_Error(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
@@ -96,8 +91,6 @@ func Test_UpdateChannelUnAuthorized_Error(t *testing.T) {
 	// Arrange
 	errorJSON := `{"message": "Invalid request"}`
 
-	ctx := t.Context()
-
 	accessToken := "access-token"
 
 	channelData := kickapitypes.UpdateChannelRequest{
@@ -113,15 +106,13 @@ func Test_UpdateChannelUnAuthorized_Error(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   mockClient,
+		HTTPClient: mockClient,
 	}
 	client, _ := kick.NewAPIClient(config)
 
 	var apiError *kickerrors.APIError
 	// Act
-	err := client.Channel().UpdateChannel(ctx, accessToken, channelData)
+	err := client.Channel().UpdateChannel(t.Context(), accessToken, channelData)
 
 	// Assert
 	if err == nil {
@@ -135,10 +126,6 @@ func Test_UpdateChannelUnAuthorized_Error(t *testing.T) {
 
 func Test_UpdateChannelWithMissingFields_Success(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
-	clientID := "test-id"
-	clientSecret := "test-secret"
-
 	accessToken := "access-token"
 
 	channelData := kickapitypes.UpdateChannelRequest{
@@ -198,16 +185,14 @@ func Test_UpdateChannelWithMissingFields_Success(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
 
 	// Act
 
-	err := client.Channel().UpdateChannel(ctx, accessToken, channelData)
+	err := client.Channel().UpdateChannel(t.Context(), accessToken, channelData)
 
 	// Assert
 
@@ -218,10 +203,6 @@ func Test_UpdateChannelWithMissingFields_Success(t *testing.T) {
 
 func Test_UpdateChannel_Success(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
-	clientID := "test-id"
-	clientSecret := "test-secret"
-
 	accessToken := "access-token"
 
 	channelData := kickapitypes.UpdateChannelRequest{
@@ -274,16 +255,14 @@ func Test_UpdateChannel_Success(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
 
 	// Act
 
-	err := client.Channel().UpdateChannel(ctx, accessToken, channelData)
+	err := client.Channel().UpdateChannel(t.Context(), accessToken, channelData)
 
 	// Assert