+++ tests/kicks/get_kicks_leaderboard_test.go
@@ -11,10 +11,6 @@ import (
 
 func TestGetKicksLeaderboardSuccess(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
-	clientID := "test-id"
-	clientSecret := "test-secret"
-
 	accessToken := "access-token"
 
 	expectedJSON := `{
@@ -64,16 +60,14 @@ func TestGetKicksLeaderboardSuccess(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
 
 	// Act
 
-	kicksLeaderboardData, err := client.Kicks().GetKicksLeaderboard(ctx, accessToken, nil)
+	kicksLeaderboardData, err := client.Kicks().GetKicksLeaderboard(t.Context(), accessToken, nil)
 
 	// Assert
 	if kicksLeaderboardData == nil {