+++ tests/users/get_users_by_id_test.go
@@ -13,23 +13,20 @@ import (
 
 func Test_GetUsersByIDMissingAccessToken_Error(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
 	httpClient := http.DefaultClient
 
 	accessToken := ""
 	users := []int64{1, 2}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 	client, _ := kick.NewAPIClient(config)
 
 	var validationError *kickerrors.ValidationError
 
 	// Act
-	usersData, err := client.User().GetUsersByID(ctx, accessToken, users)
+	usersData, err := client.User().GetUsersByID(t.Context(), accessToken, users)
 
 	// Assert
 	if usersData != nil {
@@ -53,8 +50,6 @@ func Test_GetUsersByIDUnAuthorized_Error(t *testing.T) {
 	// Arrange
 	errorJSON := `{"message": "Invalid request"}`
 
-	ctx := t.Context()
-
 	accessToken := "access-token"
 	users := []int64{1, 2}
 
@@ -65,15 +60,13 @@ func Test_GetUsersByIDUnAuthorized_Error(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     "test-id",
-		ClientSecret: "test-secret",
-		HTTPClient:   mockClient,
+		HTTPClient: mockClient,
 	}
 	client, _ := kick.NewAPIClient(config)
 
 	var apiError *kickerrors.APIError
 	// Act
-	usersData, err := client.User().GetUsersByID(ctx, accessToken, users)
+	usersData, err := client.User().GetUsersByID(t.Context(), accessToken, users)
 
 	// Assert
 	if err == nil {
@@ -91,10 +84,6 @@ func Test_GetUsersByIDUnAuthorized_Error(t *testing.T) {
 
 func Test_GetUsersByID_Success(t *testing.T) {
 	// Arrange
-	ctx := t.Context()
-	clientID := "test-id"
-	clientSecret := "test-secret"
-
 	accessToken := "access-token"
 	users := []int64{1, 2}
 
@@ -137,16 +126,14 @@ func Test_GetUsersByID_Success(t *testing.T) {
 	}
 
 	config := kickapitypes.APIClientConfig{
-		ClientID:     clientID,
-		ClientSecret: clientSecret,
-		HTTPClient:   httpClient,
+		HTTPClient: httpClient,
 	}
 
 	client, _ := kick.NewAPIClient(config)
 
 	// Act
 
-	usersData, err := client.User().GetUsersByID(ctx, accessToken, users)
+	usersData, err := client.User().GetUsersByID(t.Context(), accessToken, users)
 
 	// Assert
 	if usersData == nil {